requirements.txt
# Voice ChatBot Robot - Python Dependencies
# For NVIDIA Jetson AGX Xavier with JetPack 5.x

# AWS Bedrock Integration
boto3==1.37.38
botocore>=1.35.38

# NVIDIA Riva Speech AI
nvidia-riva-client==2.19.1

# Audio Processing
pyaudio

# RAG (Retrieval-Augmented Generation)
sentence-transformers
torch>=1.9.0  # Required by sentence-transformers
transformers>=4.0.0  # Required by sentence-transformers

# Numerical Computing
numpy

# Optional: Development/Testing
# pydub  # For audio file manipulation
# soundfile  # For audio I/O
